Move some number utils to lib-common/utils/number
