+++ frontend/src/employee-frontend/components/absences/StaffAttendance.tsx
@@ -19,7 +19,7 @@ import { StaffAttendance, StaffAttendanceGroup } from '../../types/absence'
 import { Loading, Result } from 'lib-common/api'
 import { getStaffAttendances, postStaffAttendance } from '../../api/absences'
 import { useDebounce } from 'lib-common/utils/useDebounce'
-import { formatDecimal } from '../../components/utils'
+import { formatDecimal } from 'lib-common/utils/number'
 import { faTimes } from 'lib-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import styled from 'styled-components'
