+++ frontend/src/employee-frontend/components/reports/Occupancies.tsx
@@ -25,7 +25,7 @@ import ReportDownload from '../../components/reports/ReportDownload'
 import { addDays, isAfter, isWeekend, lastDayOfMonth } from 'date-fns'
 import { formatDate } from '../../utils/date'
 import { DATE_FORMAT_ISO } from '../../constants'
-import { formatPercentage, formatDecimal } from '../../components/utils'
+import { formatPercentage, formatDecimal } from 'lib-common/utils/number'
 import { SelectOptionProps } from '../../components/common/Select'
 import { fi } from 'date-fns/locale'
 import { CareArea } from '../../types/unit'
