+++ frontend/src/employee-frontend/components/unit/tab-groups/groups/Group.tsx
@@ -44,7 +44,7 @@ import { capitalizeFirstLetter, formatName } from '../../../../utils'
 import { StatusIconContainer } from '../../../../components/common/StatusIconContainer'
 import { UnitBackupCare } from '../../../../types/child'
 import { updateBackupCare } from '../../../../api/child/backup-care'
-import { formatPercentage } from '../../../../components/utils'
+import { formatPercentage } from 'lib-common/utils/number'
 import { DataList } from '../../../../components/common/DataList'
 import { Gap } from 'lib-components/white-space'
 import IconButton from 'lib-components/atoms/buttons/IconButton'
